+++ prost-reflect/src/dynamic/serde/ser/mod.rs
@@ -52,7 +52,7 @@ impl<'a> Serialize for SerializeWrapper<'a, DynamicMessage> {
 
 fn count_dynamic_message_fields(value: &DynamicMessage, options: &SerializeOptions) -> usize {
     if options.skip_default_fields {
-        value.fields.values().filter(|v| v.is_populated()).count()
+        value.fields.values().filter(|v| v.has()).count()
     } else {
         value.fields.values().filter(|v| v.value.is_some()).count()
     }
@@ -67,7 +67,7 @@ where
     S: SerializeMap,
 {
     for field in value.fields.values() {
-        if (!options.skip_default_fields && field.value.is_some()) || field.is_populated() {
+        if (!options.skip_default_fields && field.value.is_some()) || field.has() {
             let name = if options.use_proto_field_name {
                 field.desc.name()
             } else {